Finished snippet implementation, support generation for all as types, synchonized parser and tokenizer
