+++ src/snippet.jl
@@ -1,6 +1,9 @@
+export Snippet
+
+
 struct Snippet{T <: NamedTuple}
     fn::Function
 end
 
 # I hope kwargs are internally represented as NamedTuple
-(sn::Snippet)(args...) = sn.fn(args...)
+(sn::Snippet)(; args...) = sn.fn(; args...)
