+++ test.jl
@@ -1,34 +1,20 @@
 include("./src/IonicEfus.jl")
 
 using .IonicEfus
-using .IonicEfus.Tokens
-using .IonicEfus.Parser
 using .IonicEfus.Ast
 using .IonicEfus.Gen
-using MacroTools
 
 const FILE = "test.efus"
 
 colors(; args...) = printstyled("HEllo world"; args...)
 
-println(
-    Gen.generate(
-        IonicEfus.parse_efus(
-            """
-            snippet(foo)
-              Label bar=4
-            end
-            """
-        )
-    )
-)
-
-#=
-banana = efus"""
-banana()
-  colors color=:blue bold=true
+component = efus"""
+MyLabel(children
+  (println("I am printing $children");)
 end
+
+MyLabel
+  colors
 """
 
-banana()
-=#
+println(component)
